Multiple changes: * Updated documentation. * Added functions that split the input data into the required number of jobs. * Changed the names of some internal functions. * Added a test patch to solve the problem with underestimating the running time. * The internal architecture of the system has been changed. * Now each job independently strives to load all available client cores.
*****************!!! ACHTUNG !!!*****************
This patch greatly changes the internal structure of the system. 
It is not compatible with already created jobs and old virtual 
machines. All jobs created in earlier versions of the package 
are no longer compatible with this one. This version has not 
been tested yet, minor patches are possible in the coming days.
*****************!!! ACHTUNG !!!*****************



git-svn-id: svn://svn.r-forge.r-project.org/svnroot/rboinc@38 edb9625f-4e0d-4859-8d74-9fd3b1da38cb